Added tag fighters cycling plus debugged

+++ src/fighters.h
@@ -1,13 +1,13 @@
 #ifndef FIGHTERS_H
 #define FIGHTERS_H
 
-byte fighter_index_first = -1;
-byte fighter_index_second = -1;
+int fighter_index_first = -1;
+int fighter_index_second = -1;
 
 // For tag matches
-byte fighter_index_first2 = -1;
-byte fighter_index_second2 = -1;
-byte fighter_index_first3 = -1;
-byte fighter_index_second3 = -1;
+int fighter_index_first2 = -1;
+int fighter_index_second2 = -1;
+int fighter_index_first3 = -1;
+int fighter_index_second3 = -1;
 
 #endif // FIGHTERS_H
